requirements.txt
# Core Trading Libraries
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0

# Async & WebSockets
aiohttp>=3.9.0
websockets>=12.0
asyncio-throttle>=1.0.0

# Technical Analysis
ta-lib>=0.4.28
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Logging & Monitoring
colorlog>=6.8.0

# CLI Dashboard (Matrix Mode)
colorama>=0.4.6
tabulate>=0.9.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
